Moved OtherFieldInfo, updated known_other_fields.
Moved the OtherFieldInfo class to derived_field.py, made all of its
attributes required, and so, correspondingly, added them back into
known_other_fields.yaml.
