+++ yt/frontends/enzo/fields.py
@@ -1,7 +1,7 @@
 import numpy as np
-import pydantic
 import yaml
 
+from yt.fields.derived_field import OtherFieldInfo
 from yt.fields.field_info_container import \
     FieldInfoContainer
 from yt.utilities.physical_constants import \
@@ -47,21 +47,6 @@ NODAL_FLAGS = {
     'AvgElec2': [1, 1, 0],
 }
 
-
-class OtherFieldInfo(pydantic.BaseModel):
-    """
-    Currently, every known other field has four pieces of information
-    specified: the name of the field within the frontend simulation code,
-    any aliases applicable to the field, the units, and the name of the
-    field used for display purposes in plots and the like. With this
-    set up, the aliases and display name attributes do not need to be
-    specified in the known_other_fields.yaml file if they aren't needed
-    or wanted.
-    """
-    code_name : str
-    aliases = ['',]
-    units : str
-    display_name = ''
 
 class EnzoFieldInfo(FieldInfoContainer):
     known_particle_fields = (
