+++ yt/frontends/enzo/known_other_fields.yaml
@@ -2,139 +2,220 @@
   aliases:
     - cooling_time
   units : s
+  display_name : None
 
 - code_name : Dengo_Cooling_Rate
+  aliases :
+    - None
   units : erg/g/s
+  display_name : None
 
 - code_name : Grackle_Cooling_Rate
+  aliases:
+    - None
   units : erg/s/cm**3
+  display_name : None
 
 - code_name : HI_kph
+  aliases:
+    - None
   units : 1/code_time
+  display_name : None
 
 - code_name : HeI_kph
+  aliases:
+    - None
   units : 1/code_time
+  display_name : None
 
 - code_name : HeII_kph
+  aliases:
+    - None
   units : 1/code_time
+  display_name : None
 
 - code_name : H2I_kdiss
+  aliases:
+    - None
   units : 1/code_time
+  display_name : None
 
 - code_name : HM_kph
+  aliases:
+    - None
   units : 1/code_time
+  display_name : None
 
 - code_name : H2II_kdiss
+  aliases:
+    - None
   units : 1/code_time
+  display_name : None
 
 - code_name : Bx
+  aliases:
+    - None
   units : code_magnetic
+  display_name : None
 
 - code_name : By
+  aliases:
+    - None
   units : code_magnetic
+  display_name : None
 
 - code_name : Bz
+  aliases:
+    - None
   units : code_magnetic
+  display_name : None
 
 - code_name : BxF
+  aliases:
+    - None
   units : code_magnetic
+  display_name : None
 
 - code_name : ByF
+  aliases:
+    - None
   units : code_magnetic
+  display_name : None
 
 - code_name : BzF
+  aliases:
+    - None
   units : code_magnetic
+  display_name : None
 
 - code_name : Ex
+  aliases:
+    - None
   units : code_magnetic/c
+  display_name : None
 
 - code_name : Ey
+  aliases:
+    - None
   units : code_magnetic/c
+  display_name : None
 
 - code_name : Ez
+  aliases:
+    - None
   units : code_magnetic/c
+  display_name : None
 
 - code_name : AvgElec0
+  aliases:
+    - None
   units : code_magnetic/c
+  display_name : None
 
 - code_name : AvgElec1
+  aliases:
+    - None
   units : code_magnetic/c
+  display_name : None
 
 - code_name : AvgElec2
+  aliases:
+    - None
   units : code_magnetic/c
+  display_name : None
 
 - code_name : RadAccel1
   aliases:
     - radiation_acceleration_x
   units : code_length / code_time**2
+  display_name : None
 
 - code_name : RadAccel2
   aliases:
     - radiation_acceleration_y
   units : code_length / code_time**2
+  display_name : None
 
 - code_name : RadAccel3
   aliases:
     - radiation_acceleration_z
   units : code_length / code_time**2
+  display_name : None
 
 - code_name : Dark_Matter_Density
   aliases:
     - dark_matter_density
   units : code_mass / code_length**3
+  display_name : None
 
 - code_name : Temperature
   aliases:
     - temperature
   units : K
+  display_name : None
 
 - code_name : Dust_Temperature
   aliases:
     - dust_temperature
   units : K
+  display_name : None
 
 - code_name : x-velocity
   aliases:
     - velocity_x
   units : code_velocity
+  display_name : None
 
 - code_name : y-velocity
   aliases:
     - velocity_y
   units : code_velocity
+  display_name : None
 
 - code_name : z-velocity
   aliases:
     - velocity_z
   units : code_velocity
+  display_name : None
 
 - code_name : RaySegments
   aliases:
     - ray_segments
-  units : dimensionless
+  units : None
+  display_name : None
 
 - code_name : PhotoGamma
   aliases:
     - photo_gamma
   units : eV/code_time
+  display_name : None
 
 - code_name : PotentialField
   aliases:
     - gravitational_potential
   units : code_velocity**2
+  display_name : None
 
 - code_name : Density
   aliases:
     - density
   units : code_mass / code_length**3
+  display_name : None
 
 - code_name : Metal_Density
   aliases:
     - metal_density
   units : code_mass / code_length**3
+  display_name : None
 
 - code_name : SN_Colour
+  aliases:
+  - None
   units : code_mass / code_length**3
+  display_name : None
 
 - code_name : Electron_Density
+  aliases:
+    - None
   units : code_mass / code_length**3
+  display_name : None
